fixed the cpp-code so that should now compile on windows.

+++ src/RcppExports.cpp
@@ -7,7 +7,7 @@
 using namespace Rcpp;
 
 // glm_elnet_c
-List glm_elnet_c(arma::mat x, Function pseudo_obs, arma::vec lambda, double alpha, bool intercept, double thresh, int qa_updates_max, size_t pmax, bool pmax_strict, int as_updates_max);
+List glm_elnet_c(arma::mat x, Function pseudo_obs, arma::vec lambda, double alpha, bool intercept, double thresh, int qa_updates_max, int pmax, bool pmax_strict, int as_updates_max);
 RcppExport SEXP projpred_glm_elnet_c(SEXP xSEXP, SEXP pseudo_obsSEXP, SEXP lambdaSEXP, SEXP alphaSEXP, SEXP interceptSEXP, SEXP threshSEXP, SEXP qa_updates_maxSEXP, SEXP pmaxSEXP, SEXP pmax_strictSEXP, SEXP as_updates_maxSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -19,7 +19,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< bool >::type intercept(interceptSEXP);
     Rcpp::traits::input_parameter< double >::type thresh(threshSEXP);
     Rcpp::traits::input_parameter< int >::type qa_updates_max(qa_updates_maxSEXP);
-    Rcpp::traits::input_parameter< size_t >::type pmax(pmaxSEXP);
+    Rcpp::traits::input_parameter< int >::type pmax(pmaxSEXP);
     Rcpp::traits::input_parameter< bool >::type pmax_strict(pmax_strictSEXP);
     Rcpp::traits::input_parameter< int >::type as_updates_max(as_updates_maxSEXP);
     rcpp_result_gen = Rcpp::wrap(glm_elnet_c(x, pseudo_obs, lambda, alpha, intercept, thresh, qa_updates_max, pmax, pmax_strict, as_updates_max));
